Big refactor, extract a changePanel event. Extract the GraphVisualization
